Fix build error. Separate gatsby and storybook.
