+++ src/components/Home/index.tsx
@@ -36,12 +36,14 @@ export const HomePageTemplate: FunctionComponent<HomePageProps> = ({
       <section id="offerings">
         <Offerings offerings={offerings.blurbs} />
       </section>
-      <CallToAction
-        headline={contactHeadline}
-        description={contactDescription}
-        url={contactEmail}
-        buttonLabel={contactButton}
-      />
+      <section id="contact">
+        <CallToAction
+          headline={contactHeadline}
+          description={contactDescription}
+          url={contactEmail}
+          buttonLabel={contactButton}
+        />
+      </section>
     </>
   </div>
 )
