+++ src/components/Link/index.tsx
@@ -1,0 +1,24 @@
+import React, { FunctionComponent } from 'react'
+
+export interface LinkProps {
+  className?: string
+  to: string
+  title?: string
+}
+
+export const SafeExternalLink: FunctionComponent<LinkProps> = ({
+  className,
+  children,
+  to,
+  title,
+}: any) => (
+  <a
+    className={className}
+    title={title}
+    href={to}
+    target="_blank"
+    rel="noopener noreferrer"
+  >
+    {children}
+  </a>
+)
