+++ src/components/layout.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react'
 import Helmet from 'react-helmet'
 import { Navbar } from './Navbar'
-import Footer from './Footer'
+import { Footer } from './Footer'
 import config from '../../config'
 import styledNormalize from 'styled-normalize'
 import { ThemeProvider, injectGlobal } from 'styled-components'
@@ -12,12 +12,12 @@ injectGlobal`
   ${styledNormalize}
 
   @import url('https://fonts.googleapis.com/css?family=Open+Sans:400,400italic,600,600italic,800,800italic');
-  
+
   * {
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
   }
-  
+
   *, .row, :after, :before {
     -moz-box-sizing: border-box;
     -webkit-box-sizing: border-box;
@@ -29,12 +29,12 @@ injectGlobal`
     font-weight: 400;
     letter-spacing: .075em;
     margin: 0;
-    
+
     ${media.greaterThan('sm')`
       font-size: 12pt;
     `}
   }
-  
+
   body {
     background: rgb(46, 56, 66);
     color: #ffffff;
@@ -43,18 +43,18 @@ injectGlobal`
     line-height: 1.65em;
     margin: 0;
     }
-  
+
   blockquote {
     border-left: solid 4px #dfdfdf;
     font-style: italic;
     margin: 0 0 2em 0;
     padding: 0.5em 0 0.5em 2em;
-    
+
     > p {
       margin: 0 !important;
     }
   }
-  
+
 `
 
 const hardforkTheme = {
